Extract pick_download_location(), calculate_required_space() utilities
We'll need the functionality of picking a download location and
calculating required space for Flatpak installation as well as for
DNF installation, so move the utility functions into
payloads/base/utils.py.

Resolves: RHEL-75804
