+++ pyanaconda/modules/payloads/payload/dnf/installation.py
@@ -36,6 +36,7 @@ from pyanaconda.modules.common.errors.installation import (
 )
 from pyanaconda.modules.common.structures.packages import PackagesConfigurationData
 from pyanaconda.modules.common.task import Task
+from pyanaconda.modules.payloads.base.utils import pick_download_location
 from pyanaconda.modules.payloads.payload.dnf.requirements import (
     apply_requirements,
     collect_dnf_requirements,
@@ -46,13 +47,14 @@ from pyanaconda.modules.payloads.payload.dnf.requirements import (
 )
 from pyanaconda.modules.payloads.payload.dnf.utils import (
     get_kernel_version_list,
-    pick_download_location,
 )
 from pyanaconda.modules.payloads.payload.dnf.validation import (
     CheckPackagesSelectionTask,
 )
 
 log = get_module_logger(__name__)
+
+DNF_PACKAGE_CACHE_DIR_SUFFIX = 'dnf.package.cache'
 
 
 class SetRPMMacrosTask(Task):
@@ -191,7 +193,9 @@ class PrepareDownloadLocationTask(Task):
 
         :return: a path of the download location
         """
-        path = pick_download_location(self._dnf_manager)
+        path = pick_download_location(self._dnf_manager.get_download_size(),
+                                      self._dnf_manager.get_installation_size(),
+                                      DNF_PACKAGE_CACHE_DIR_SUFFIX)
 
         if os.path.exists(path):
             log.info("Removing existing package download location: %s", path)
